[refactor] Move connectivity out of swarm.js
The DHT's readiness and NAT verdict, the canary probe, the liveness poll
and the debounced status frame — 610 lines that read the swarm handle and
report. Its sixteen counters and six timers were the largest group
destroySwarm cleared by hand; the connection layer keeps three note* calls.
